+++ src/features/tasks/TasksPage/Search/index.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { useQueryParameter, useReplaceQueryParameter } from "../../queryParamaters";
+import { useQueryParameter, useReplaceQueryParameter } from "../../../../common/utils/queryParamaters";
 import Input from "../Input";
 import searchQueryParamName from "../searchQueryParamName";
 import { Wrapper } from "./styled";
